Implement support for user pronunciation dictionary.
A dictionary is a text file in UTF-8. Entries are separated by new
lines. A word is separated from the pronunciation by spaces. The
pronunciation must include only Russian letters and + or - symbols.
Stress is marked by + before a vowel. If there are more then one word
in the pronunciation, they are separated by -.