Start reimplementing user dictionary support
As we now have our own ini parser, we can add more features to the
dictionary implementation, so change the dictionary format. Stop using
sorted arrays to store a dictionary and switch to a data structure
which is more convinient and is traditionally used to implement
associative arrays, namely a red-black tree. Use an implementation of
a red-black tree in C by Niels Provos, which seems to be a classic
choice for C programs. The only improvement compared to our old user
dictionary implementation at this point is support for multipple
dictionaries, so a user can keep their personal dictionary and use
dictionaries created by other people.