Reconsider the way we are searching for the user settings
On initialization accept the path to the configuration directory
instead of just the path to the configuration file. Inside this
directory look for the configuration file with the predifined name.
Assume that all the user dictionaries are in the dicts subdirectory
and try to parse each file in this directory as a dictionary.