Improve handling of user dictionaries and variants
Unify the formats, so that a variant is just a named dictionary. If
the left-hand side of an entry is a single string, consider it a whole
word. Otherwise, assume that the first word is the prefix, the second
word is the substring which will be replaced with the right-hand side
of the entry, and the third word is the suffix. Prefixes and suffixes
are regular expressions.