Try to fix a problem with speeding up speech
When we increased speech rate without using libsonic, sometimes we
asked HTS Engine to make a sound shorter than it is actually possible,
and you could hear this skips in the resulting speech. Try to prevent
this by not asking the engine to make phones less than nstate frames long.