A tiny improvement of the sentence selection script
If a phoneme is mostly represented by a particular letter combination,
then maximizing bigram coverage instead of diphone coverage might
produce too few examples. Of cause, it would be better to use phonemic
transcriptions of the sentences, but if we can't, at least treat these
combinations specially, that as as single letters, so we can pick up
more examples of them in different contexts.