NVDA synthDriver: do not use voice pairs to support automatic language switching
The list of voices now contains only actual voices, and the driver now
supports the variant setting, which allows a user to select a second
voice for automatic switching between languages. We don't restrict the
language of the synthesizer to the language of NVDA anymore, and the
language of the current voice is reported to the screen reader.