Add support for alternative implementations of an HTS engine
All implementations derive from an abstract base class. They are only
responsible for the synthesis itself, unlike the old implementation,
which included a lot of unrelated code. The steps which don't depend
on the engine are represented by separate classes. So, the user can
now choose between two implementations: the standard one using the HTS
engine developed by the HTS working group, and the implementation
using Mage for parameter generation, which should respond more quickly
on long sentences.
In the process of this rewrite we have changed the way possible
post-processing can be applied to the synthesized speech before it is
passed to the client. Each post-processing action is represented by a
class and they are organized in a chain. At run-time we decide which
post-processors should be included in the chain. So it is possible to
plug in such things as filters, if the user requests it.