Language and voice data does not have to be stored in one place any more
As such data may now be distributed in separate packages, we should
take into account situations when we cannot install all the files in
the same directory. For example, NVDA extracts each addon into its own
directory. So it is now possible to provide a list of directories
where language and voice data can be found. On initialization RHVoice
will add those languages and voices to the corresponding lists.