Add a command to create low-pass filter and the corresponding decision tree
We use fixed value of mvf the way standard HTS voices do, but add an
exception for sibilants, setting mvf to 2 khz. This is a
compromise which tries to remove most unpleasant instances of
buzzy quality from synthesized speech without switching to dynamic
mvf, which may not always be extracted correctly. At least this solves
the problem with Anna, where synthesizing the zh phone using the
default value of mvf gives really unacceptable results.

We make this a separate command, so it is possible to modify the tree
by hand and rerun the generation of the voice data file.