Android: fix downloads on Android<5.0
Bintray now supports only TLS v1.2, which broke voice downloads on
Android<5.0. Google recommends using Google Play services to update
the security provider. But users will not be able to install the
voices on those old devices if Google Play Services are not installed
or up-to-date. For now I will bundle the Conscrypt library with the
app, though it makes the app larger.