Android: add the ability to listen to an example of synthesized speech
If the voice is not installed, a user can listen to a pre-recorded
sample. The user doesn't have to download the voice to find out if they
like how it sounds. If the voice is already installed, we make it say
a pre-defined message to show that it is working.