User dictionary: support for being included into the language pack
Trusted users can maintain dictionaries for their languages and
contribute the updates. User's won't need to install such dictionaries separately.
The plan is to implement fast-loading binary format, but there is no
time to work on it right now, so for now just load the text files,
only they will be searched both in the package in in the user's
configuration directory.