Revert "Android: replace edit fields with sliders for volume and rate settings"
This reverts commit abe43705467ce8cb67dacb9909c8bf6f177edd14.

I didn't check earlier, but sliders are more problematic with TalkBack
on older versions of Android: the change is too large and TalkBack
doesn't allow to enter the value manually. For now undo the change,
because I have no time to figure it out right now. Maybe later make
changes to use sliders only on later versions of Android (first
determine from which version to start).