Disconnect conditional building of speech_dispatcher module and server (#362)
Instead they are now built depending solely on the BUILD_ options, but these options are initialized depending on the presence of components' dependencies. So a user can override these options, if he really needs so (it would fail, if the dependencies are not met, but he can provide them too via the cache vars set by pkg-config).