A reference tree for duration.
This is to always use the "is this a pause?" question as the root question of the duration tree. In the default setup sometimes some pauses may become a part of the cluster at the leaf node and cause durations of other sounds at that node to be too long. So we separate silences from everything else here right away.