Another kind of simple language switching. (#696)
This is meant for the case of a second language sharing most of the alphabet with the first language. The first language is the default. The second language is chosen only if a matching word is found with no clashes with the first language. The phoneme mapping is used the same way as for pseudo-English.