Material 3 and edge-to-edge (#952)
* Update gradle to 8.14

* Update AGP to 8.9.2

- Added multidex library since it is no longer shipped with AGP

- Added additional proguard rules because R8 full mode was enabled by default in a previous release

* Update compile sdk and build tools to 35

* Use material 3 components

- Changed the application theme to Theme.Material3.DayNight
- Changed the play/install Buttons in the voice submenu to MaterialButtons
- Changed the ProgressBar in the voice submenu to a CircularProgressIndicator

* Implement edge-to-edge

In order to better accommodate this change, I needed to do a few things first. - The main layout was abstracted into a ToolbarFragment so the MaterialToolbar's lift on scroll functionality would work best (I've found that it frequently has issues with anything that isn't a recycler view and if you swap between multiple different scrolling lists with a single toolbar). The ToolbarFragment handles the initial basic setup of the screen and gives you access to the toolbar and the ability to replace the frame layout inside. This allows it to be reused for every screen in the app. - As a result of the previous point, I converted the language list into a recycler view. - InsetUtil was created to easily add inset listeners with the max insets created on each side of a view based on the system bar and display cutout insets.

The app now displays properly with edge to edge enabled while using the shortEdges window mode which allows the UI to appear around display cutout insets when in certain orientations.

* Apply dynamic colors to activities at startup

* Improve font and icon visibility
Font size for the main list items was increased to 18sp and icon size for the voice list items was increased to 48dp to more closely match the original look